Breaking PixelToLatLon into its own area.
